refactor: moved project/route into project/[slug]/route
